Add files via upload
In this work, three years worth of data (2012, 2013, 2014) from Lending Club -the online platform for personal lending- are used to train and test five statistical learning algorithms: kNN, Logistic Regression, Rpart, Decision Tree, C5.0, and Random Forest. 
The goal is to predict if the loans will be paid or not. Every model is evaluated for accuracy, and the best model is used to classify 10% of the loans made during 2015.